+++ PlasmaSwiftLibTests/BlockTests.swift
@@ -12,5 +12,23 @@ import BigInt
 @testable import PlasmaSwiftLib
 
 class BlockTests: XCTestCase {
-
+    let testHelpers = TestHelpers()
+    
+    func testBlockHeader() {
+//        guard let blockHeader1 = testHelpers.formBlockHeaderForBlock() else {return}
+//        let data = blockHeader1.data
+//        guard let blockHeader2 = BlockHeader(data: data) else {return}
+//        print("block header passed")
+//        XCTAssert(blockHeader1 == blockHeader2)
+    }
+    
+    func testBlock() {
+//        guard let blockHeader = testHelpers.formBlockHeaderForBlock() else {return}
+//        guard let signedTransactions = testHelpers.formSignedTransactionsForBlock() else {return}
+//        guard let block1 = Block(blockHeader: blockHeader, signedTransactions: signedTransactions) else {return}
+//        let data = block1.data
+//        guard let block2 = Block(data: data) else {return}
+//        print("block passed")
+//        XCTAssert(block1 == block2)
+    }
 }
